test: Move `testRenderExceptionSetsDisplayErrorsInDebugMode` from `ApplicationErrorHandlerTest` to `ApplicationTest` for better organization..
